Refactored slider
